Most exception added

+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/grocerystore/service/GroceryStoreService.java
@@ -501,11 +501,17 @@ public class GroceryStoreService {
 	
 	@Transactional
 	public PerishableItem getPerishableItemsByID(Integer id) {
-
-		return perishableItemRepository.findByItemID(id);
+		PerishableItem pitem = perishableItemRepository.findByItemID(id);
+		
+		if (pitem == null) 
+			throw new IllegalArgumentException("No such perishable item. Please search by another ID.");
+		
+		return pitem;
 	}
 	@Transactional
 	public PerishableItem deletePerishableItems(PerishableItem pitem) {
+		
+		if (pitem == null) throw new IllegalArgumentException("Please enter an item to delete.");
 		perishableItemRepository.delete(pitem);
 		return pitem;
 		
@@ -513,10 +519,37 @@ public class GroceryStoreService {
 	
 	@Transactional
 	public List<PerishableItem> getPerishableItemsByProductName(String name) {
-		return perishableItemRepository.findByProductName(name);
+		List<PerishableItem> pitem = perishableItemRepository.findByProductName(name);
+		if (pitem == null) 
+			throw new IllegalArgumentException("No such perishable items. Please search by another name.");
+		
+		return pitem;
 	}
 	@Transactional
 	public PerishableItem updatePerishableItem(PerishableItem pitem, String productName,Float price, Boolean availableOnline,Integer numInStock, Integer pointPerItem) {
+		
+		ArrayList<String> errors = new ArrayList<String>();
+		
+		if (pitem == null) throw new IllegalArgumentException("Please enter an item to update.");
+		
+		if (productName == null || productName.trim().length() == 0) {
+			errors.add("Item name is empty!");
+		}
+		if (price == null) {
+			errors.add("Price is empty!");
+		}
+		if (availableOnline == null) {
+			errors.add("Please state whether this item is available online!");
+		}
+		if (numInStock == null) {
+			errors.add("Please state the amount of stock!");
+		}
+		if (pointPerItem == null) {
+			errors.add("Please state the amount of point given per item!");
+		}
+		String listErrors = String.join(", ", errors);
+		if(errors.size() != 0) throw new IllegalArgumentException("Item unchanged. " + listErrors);
+		
 		pitem.setProductName(productName);
 		pitem.setPrice(price);
 		pitem.setAvailableOnline(availableOnline);
@@ -573,22 +606,54 @@ public class GroceryStoreService {
 	
 	@Transactional
 	public NonPerishableItem getNonPerishableItemsByID(Integer id) {
+		NonPerishableItem npitem = nonPerishableItemRepository.findByItemID(id);
+		if (npitem == null) 
+			throw new IllegalArgumentException("No such non perishable item. Please search by another ID.");
 		
 		return nonPerishableItemRepository.findByItemID(id);
 	}
 	
 	@Transactional
 	public List<NonPerishableItem> getNonPerishableItemsByProductName(String name) {
-		return nonPerishableItemRepository.findByProductName(name);
+		
+		List<NonPerishableItem> npitem = nonPerishableItemRepository.findByProductName(name);
+		if (npitem == null) 
+			throw new IllegalArgumentException("No such non perishable items. Please search by another name.");
+		
+		return npitem;
 	}
 	
 	@Transactional
 	public NonPerishableItem deleteNonPerishableItems(NonPerishableItem npitem) {
+		if (npitem == null) throw new IllegalArgumentException("Please enter an item to delete.");
 		nonPerishableItemRepository.delete(npitem);
 		return npitem;
 	}
 	@Transactional
 	public NonPerishableItem updateNonPerishableItem(NonPerishableItem npitem, String productName,Float price, Boolean availableOnline,Integer numInStock, Integer pointPerItem) {
+		
+ArrayList<String> errors = new ArrayList<String>();
+		
+		if (npitem == null) throw new IllegalArgumentException("Please enter an item to update.");
+		
+		if (productName == null || productName.trim().length() == 0) {
+			errors.add("Item name is empty!");
+		}
+		if (price == null) {
+			errors.add("Price is empty!");
+		}
+		if (availableOnline == null) {
+			errors.add("Please state whether this item is available online!");
+		}
+		if (numInStock == null) {
+			errors.add("Please state the amount of stock!");
+		}
+		if (pointPerItem == null) {
+			errors.add("Please state the amount of point given per item!");
+		}
+		String listErrors = String.join(", ", errors);
+		if(errors.size() != 0) throw new IllegalArgumentException("Item unchanged. " + listErrors);
+		
 		npitem.setProductName(productName);
 		npitem.setPrice(price);
 		npitem.setAvailableOnline(availableOnline);
